Reads a text file and prints it to POSIX stdout.

+++ 0x15-file_io/0-read_textfile.c
@@ -11,7 +11,7 @@
  */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	int fd, bytes_read, bytes_written;
+	ssize_t fd, bytes_read, bytes_written;
 	char *buffer;
 
 	if (filename == NULL)
